Setup up bundler and the rake tests for travis-ci.
The dependencies specified in the Gemfile are probably a little
overly-specific, I just wanted to come up with something that
would work first.

The multiruby task that was taken out of test.rake can be
replaced by using travis-ci to test multiple versions of ruby.
I'll configure it that way once I get ruby 1.9.2 to pass all
the tests.