Improve line numbers CLI options
* Make the short form have the positive meaning
* Make the long form have the meaning suggested by its name
* Add negative of the long form
* Add a trailing s to the long form, which makes more sense

This means the short form usually does nothing. On the positive side, it has the
meaning it had from the beginning. Also, once command line arguments can be set
in config files, it can be used to override those. We can add a new short option
to be the short version of --no-line-numbers.