Refactor CLI commands.
Right now we have 2 commands, YamlCommand and ReekCommand.
This is bad for 2 reasons:
1.) It's highly confusing since it actually implies that those 2 commands are
doing something different when in fact they are doing almost the exact
same thing.
2.) The point in which they differ is the output format which is
something they shouldn't even know about. The output format should be
part of the reports, not of the commands.

After this refactoring we'll only have one command 'ReekCommand', which
knows nothing about the output format and instead delegates this
responsibility to our reporter.