Fix file arguments without TTY
When using reek as part of a build process, there might not be a TTY.
This does not necessarily mean Ruby code is piped into reek via $stdin,
but reek ignores file arguments in this case. This leads to unexpected
behavior as the build script will continue to work on the terminal, but
will not scan any files on the build server.

This is fixed by only reading from $stdin when no file arguments are
given.