Use Aruba for the cucumber scenarios
This replaces our home-grown method of running reek in the scenario's
with Aruba's methods. One of the great advantages is that each scenario
will run in its own separate environment with a working directory that
starts empty and is cleaned up at the end.

Because of this change, it is now possible to test all config file
finder scenarios in temporary directories instead of pulluting the
developer's home and project directories.

Many of the sample files are no longer needed because they are now
generated in a temporary directory. These have been removed. Some other
sample files are kept because they are still used in the specs.