Silence factory_girl warnings
`factory_girl` gem outputs a couple ruby warnings. Unfortunately, at the
time of this commit this is not yet fixed in `factory_girl` repo. We
have to wait till the gem is updated and a new version released.

Also, the gem version assertion is added so we remember to remove
silencer once `factory_girl` gem is (finally) updated.