Handle modifier style use of module_function for UtilityFunction
To be able to track calls to module_function in all cases, these calls
are handled during the walking of the tree by TreeWalker. Parent
contexts track the currently set visibility and pass it on to any new
child contexts. If module_function is used with arguments, the parent
context will instead modify the visibility of existing children.

In order to have the final visibility at hand while detecting smells,
the smell detection is done after the entire tree has been walked.