Add Code Climate JSON report format
This commit introduces a new formatter intended for use with Code
Climate's engine platform.

https://github.com/codeclimate/spec

Further integration work will be required, but the aim is that the Reek
repository will act as the home for Code Climate's Reek engine, rather
than having a separate repository for the engine.

The core team have discussed and agreed to this approach.

cc @mr_b [CHECK]