Use AST::Node's own stored source as a base for formatting
AST::Node stores the parsed source so we can use that directly. As a
result, some smell masking may have to be adjusted.

Since SexpFormatter is no longer needed, it is removed, as is its
dependency, Unparser.

Since Parser 2.3.0 behaves differently on certain syntax errors and support
for it is part of an upcomping change, the allowed versions of Parser
are limited, as Unparser did in the past.