Make mutant task more robust
Specifying 'master^' makes sure that in most day-to-day cases, the
mutant task will find some commits to check. The is preferable to
checking for the current branch, which seems to fail on Travis.

An alternative is to use HEAD~n, for some value of n. However, this
could hide problems while a PR is being developed, while potentially
prolonging build failures on master after merging a branch that mutant
does not approve of.