Emit fingerprints in Code Climate reporter
This enabled Code Climate to more accurately identify whether an issue
is new, as opposed to changed.

I chose to look at the `warning.parameters` value rather than the
`warning.message` value, because the message (for some smell types) will
change over time. For example:

If you have this code:

```ruby
class Alfa
  Bravo = Charlie = Delta = 1
end
```

You will have a warning with message "Alfa has 3 constants".

Later, when you add a 4th constant, it will change to "Alfa has 4
constants".

Ideally, this would be flagged as the same Code Climate issue, but with
a greater remediation points value, so we can flag it as getting worse.
For now, I think it's good to simply treat it as the same issue.

The way this is accomplished is by CodeClimateFingerprint maintaining a
list of parameters to ignore (:count and :depth, currently). I think
this is a good light-weight approach, but am open to the idea of moving
it somewhere else if you're concerend that may fall out-of-sync over
time.

There are two smell types which don't seem to be identifiable based on
smell_type + context + parameters, which are ManualDispatch and
NilCheck. I believe those could become identifiable if they had
parameters, but haven't done that.