Implement `--force-exclusion` flag
Fixes #1196

This introduces a --force-exclusion flag that mirrors the behavior of
Rubocop's similarly-named flag. If the config contains an exclude_path
for smelly_path and CLI is called with

reek ./smelly_path/file.rb

Then the behavior of Reek is unchanged and the file will be checked for smell violations

But if the CLI is called with

reek --force-exclusion ./smelly_path/file.rb

Then the file will not be checked and Reek will exit with 0.

Notes:

- For the Source Locator to know about the option being set we have to
pass the options in. I'm following the existing pattern here of named
arguments with a default.

- Choosing to ignore a file is different than ignoring a path, so there's a new
method encapsulate it. If the force_exclusion option is in effect, the
`ignore_file?' method looks to see if any of the file's parent paths are
excluded.

- Even though I changed code in `reek/cli/application` I did not add
tests. I didn't see that there were any existing tests that cared how an
instance of `Source::SourceLocator` was constructed (by mocking the
call, or similar approaches). Since the tests of application continued to pass, I
didn't see a reason to add new tests here.

- The `ascend` method wasn't fully enumerable before Ruby 2.3
This commit uses the version of ascend in place since Ruby 1.9 for
backwards compatibility with Ruby 2.1 and 2.2.

Background:

The 1.9.3 - 2.2 version of the docs: http://ruby-doc.org/stdlib-1.9.3/libdoc/pathname/rdoc/Pathname.html#method-i-ascend

The 2.3 and up version http://ruby-doc.org/stdlib-2.3.1/libdoc/pathname/rdoc/Pathname.html#method-i-ascend

Note the "Returns an Enumerator if no block was given" section that's
been added.