Properly handle absolute paths
Fixes #1202

I came at this a few different ways. First I looked at expanding all
paths when creating a new SourceLocator. That worked, but made a bunch
of tests fail because they were expecting output that contained relative
paths.

So, to fix the bug and not force a rewrite of all the tests I settled on
doing the conversion within the path_excluded method. Doing it here
ensures we're comparing two absolute paths and it prevents the expanded
paths from leaking throughout the app.

I tried implementing a feature test for this but had no success. It's
weirdly hard to get an absolute path for a file you create through
Aruba. But I'm not that familiar with Cucumber and Aruba, so maybe I'm
missing something.