Fixes cases where modules containing only method: thing don't reek
"Namespace modules" are allowed without documentation. However, reek
incorrectly thinks a module containing method calls like `some_method: thing`
is a namespace module. To fix this, it was suggested in #1227 that if a
module contains *anything* other than a casgn, class or module def, then
it should not be considered a namespace module.

Addresses #1227