Initialize a fresh detector for each context
The existing implementation used the same detector for all contexts,
which resulted in constantly passing around the current context and
other values as parameters. This in turn has lead to a large number of
cases of feature envy and similar code smells.

By using each detector instance for just one context, we can migrate
to a design where more of the state of the ongoing detection is stored
in the detector and less is passed around.