Use instance attributes in detector implementations
This is the logical follow-up of pull request #1234. There, a separate
detector instance per context was introduced, to make context available
as an attribute in the detector. Now, this attribute is used in the
sniff implementations and the old parameter is dropped. This cleans up
some code smells as well.

Some detectors can probably be simplified further, but I have mostly
tried to keep the structure as is for now.