Handle encoding and other errors when parsing the source
Handles any errors in SourceCode the same way as other errors in the
examination process by wrapping it in the same error handling logic as
the smell detection. This results in either the message being printed,
or the exception being re-raised, in wrapped form if necessary.

For encoding errors, a special EncodingError wrapper class is added.