Add support for --stdin-filename
When other tools, like automatic linters within editors, run a sample
through reek, they typically do so via stdin. However, if there are reek
config exceptions to apply to that file, then reek has no way of knowing
what the original filename was, triggering false positives.

This adds an `--stdin-filename` CLI option that takes the
original filename, and passes it through to the Examiner and SourceCode
file to use as the `origin` attribute, allowing the configuration
directives to apply.