Set link metadata in gemspec
It is no longer possible to set these via the rubygems.org interface.
Instead, these should now be set in the gemspec. Wiki and mailing list
are omitted since the former is disabled and the latter hasn't seen
activity since 2012.

See https://blog.rubygems.org/2019/03/08/and-then-there-was-one-metadata-links.html
for more info on the metadata UI removal.