Rename Code Climate related classes
The Code Climate classes were living under the Report namespace but did
not all implement report classes. With this change, they are moved to
the CodeClimate namespace instead. They are also moved to the
lib/reek/code_climate directory to put their file names in line with
their namespace name.

The spec files are moved as well, fixing an outstanding
RSpec/SpecFilePathFormat offense.