Fix Reek installation issue when Yard is installed
As per #1627, Yard can't process Reek because `docs/yarn_plugin.rb` is
missing. It's not included in the Gem, so this PR adds it to correct
this error.

This was reported upstream in
https://github.com/lsegal/yard/issues/1462, but it seems that it either
wasn't, or people were happy enough with the workaround (touching the
missing file in the Gem's installed location). The workaround doesn't
work when using `asdf`/`mise` with [default
Gems](https://mise.jdx.dev/lang/ruby.html#default-gems). If Yard's automatic
documentation is enabled (`yard config --gem-install-yri`) then the Ruby
build fails.